Move remaining files and updated init
