+++ deeplabcut/__init__.py
@@ -26,36 +26,32 @@ from deeplabcut import DEBUG
 # DLClight version does not support GUIs. Importing accordingly
 import matplotlib as mpl
 
-if os.environ.get("DLClight", default=False) == "True":
+try:
+    import wx
+    mpl.use("WxAgg")
+    from deeplabcut import generate_training_dataset
+    from deeplabcut import refine_training_dataset
+    from deeplabcut.generate_training_dataset import (
+        dropannotationfileentriesduetodeletedimages,
+        comparevideolistsanddatafolders,
+        dropimagesduetolackofannotation,
+        adddatasetstovideolistandviceversa,
+        dropduplicatesinannotatinfiles,
+    )
+    from deeplabcut.gui.launch_script import launch_dlc
+    from deeplabcut.gui.label_frames import label_frames
+    from deeplabcut.gui.refine_labels import refine_labels
+
+    from deeplabcut.utils import select_crop_parameters
+    from deeplabcut.utils.skeleton import SkeletonBuilder
+    from deeplabcut.refine_training_dataset.tracklets import refine_tracklets
+except ModuleNotFoundError:
     print(
         "DLC loaded in light mode; you cannot use any GUI (labeling, relabeling and standalone GUI)"
     )
     mpl.use(
         "AGG"
     )  # anti-grain geometry engine #https://matplotlib.org/faq/usage_faq.html
-else:  # standard use [wxpython supported]
-    mpl.use("WxAgg")
-    from deeplabcut import generate_training_dataset
-    from deeplabcut import refine_training_dataset
-    from deeplabcut.generate_training_dataset import (
-        label_frames,
-        dropannotationfileentriesduetodeletedimages,
-        comparevideolistsanddatafolders,
-        dropimagesduetolackofannotation,
-    )
-    from deeplabcut.generate_training_dataset import (
-        multiple_individuals_labeling_toolbox,
-    )
-    from deeplabcut.generate_training_dataset import (
-        adddatasetstovideolistandviceversa,
-        dropduplicatesinannotatinfiles,
-    )
-    from deeplabcut.gui.launch_script import launch_dlc
-
-    from deeplabcut.refine_training_dataset import refine_labels
-    from deeplabcut.utils import select_crop_parameters
-    from deeplabcut.utils.skeleton import SkeletonBuilder
-    from deeplabcut.refine_training_dataset.tracklets import refine_tracklets
 
 
 from deeplabcut.create_project import (
@@ -63,34 +59,31 @@ from deeplabcut.create_project import (
     create_new_project_3d,
     add_new_videos,
     load_demo_data,
-)
-from deeplabcut.create_project import (
     create_pretrained_project,
     create_pretrained_human_project,
 )
-from deeplabcut.generate_training_dataset import extract_frames
 from deeplabcut.generate_training_dataset import (
     check_labels,
     create_training_dataset,
+    extract_frames,
     mergeandsplit,
 )
 from deeplabcut.generate_training_dataset import (
     create_training_model_comparison,
+    create_multianimaltraining_dataset,
     cropimagesandlabels,
 )
-from deeplabcut.generate_training_dataset import create_multianimaltraining_dataset
 from deeplabcut.utils import (
     create_labeled_video,
     create_video_with_all_detections,
     plot_trajectories,
     auxiliaryfunctions,
-)
-from deeplabcut.utils import (
     convert2_maDLC,
     convertcsv2h5,
     convertannotationdata_fromwindows2unixstyle,
+    analyze_videos_converth5_to_csv,
+    auxfun_videos,
 )
-from deeplabcut.utils import analyze_videos_converth5_to_csv, auxfun_videos
 
 from deeplabcut.utils.auxfun_videos import ShortenVideo, DownSampleVideo, CropVideo
 
@@ -99,26 +92,19 @@ from deeplabcut.utils.auxfun_videos import ShortenVideo, DownSampleVideo, CropVi
 from deeplabcut.pose_estimation_tensorflow import (
     train_network,
     return_train_network_path,
-)
-from deeplabcut.pose_estimation_tensorflow import (
     evaluate_network,
     return_evaluate_network_data,
     evaluate_multianimal_crossvalidate,
-)
-
-from deeplabcut.pose_estimation_tensorflow import (
     analyze_videos,
     analyze_time_lapse_frames,
     convert_detections2tracklets,
-)
-from deeplabcut.pose_estimation_tensorflow import (
     extract_maps,
     visualize_scoremaps,
     visualize_locrefs,
     visualize_paf,
     extract_save_all_maps,
+    export_model,
 )
-from deeplabcut.pose_estimation_tensorflow import export_model
 
 from deeplabcut.pose_estimation_3d import (
     calibrate_cameras,
