+++ deeplabcut/generate_training_dataset/__init__.py
@@ -8,17 +8,6 @@ https://github.com/AlexEMG/DeepLabCut/blob/master/AUTHORS
 Licensed under GNU Lesser General Public License v3.0
 """
 
-import os
-
-if os.environ.get("DLClight", default=False) == "True":
-    # print("DLC loaded in light mode; you cannot use the labeling GUI!")
-    pass
-else:
-    from deeplabcut.generate_training_dataset.auxfun_drag_label import *
-    from deeplabcut.generate_training_dataset.labeling_toolbox import *
-    from deeplabcut.generate_training_dataset.multiple_individuals_labeling_toolbox import *
-    from deeplabcut.generate_training_dataset.frame_extraction_toolbox import *
-
 from deeplabcut.generate_training_dataset.frame_extraction import *
 from deeplabcut.generate_training_dataset.trainingsetmanipulation import *
 from deeplabcut.generate_training_dataset.multiple_individuals_trainingsetmanipulation import *
