+++ deeplabcut/refine_training_dataset/__init__.py
@@ -8,14 +8,6 @@ https://github.com/AlexEMG/DeepLabCut/blob/master/AUTHORS
 Licensed under GNU Lesser General Public License v3.0
 """
 
-import os
 
-if os.environ.get("DLClight", default=False) == "True":
-    # print("DLC loaded in light mode; you cannot use the relabeling GUI!")
-    pass
-else:
-    from deeplabcut.refine_training_dataset.refinement import *
-    from deeplabcut.refine_training_dataset.auxfun_drag import *
-    from deeplabcut.refine_training_dataset.tracklets import *
-
+from deeplabcut.refine_training_dataset.tracklets import *
 from deeplabcut.refine_training_dataset.outlier_frames import *
